some cleanup & restructuring
