+++ packages/sdk/src/library/wallet.ts
@@ -13,7 +13,7 @@
 import { Keyring } from '@polkadot/api';
 import { mnemonicGenerate } from '@polkadot/util-crypto';
 import { Mnemonic, ethers } from 'ethers';
-import { deferTask } from '../utils/deferTask';
+import { deferTask } from './utils';
 import { getAutoIdFromSeed, getIdentityFromSeed } from './did';
 
 /**
