--- packages/sdk/src/utils/deferTask.ts
@@ -1,3 +1,0 @@
-export const deferTask = <T>(task: () => T) => {
-  return new Promise<T>((resolve) => setTimeout(() => resolve(task())));
-};
